commit - 20220524 - kosmo - Proj01

+++ src/main/java/ex01/DeleteController.java
@@ -14,6 +14,9 @@ import javax.servlet.http.HttpServletResponse;
 @WebServlet("/member/delete.do")
 public class DeleteController extends HttpServlet {
 	
+	
+	MemberDAO dao = new MemberDAO();
+	
 	public static void alertLocation (HttpServletResponse resp, String msg, String url) {
 		try {
 			resp.setContentType("text/html;charset=UTF-8");
@@ -30,23 +33,6 @@ public class DeleteController extends HttpServlet {
 		} catch(Exception e) {
 		}
 	}
-	
-	public static void alertBack (HttpServletResponse resp, String msg ) {
-		try {
-			resp.setContentType("text/html;charset=UTF-8");
-			PrintWriter writer = resp.getWriter();
-			
-			String script = ""
-						  + "<script>"
-						  + "    alert('" + msg + "');"
-						  + "    history.back();"
-						  + "</script>";
-			writer.print(script);
-			
-		} catch (Exception e) {
-			
-		}
-	}
 
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
@@ -58,19 +44,18 @@ public class DeleteController extends HttpServlet {
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		// post
 		String id = req.getParameter("id");
-		String pwd = req.getParameter("pwd");
 		
-		MemberDAO dao = new MemberDAO();
-		boolean confirmed = dao.confirmPassword(pwd, id);
+		int result = dao.deletePost(id);
 		dao.close();
 		
+		if(result == 1) {
+			System.out.println("삭제 성공공");
+			dao.deletePost(id);	//게시물 삭제
+		}else if (result == 0) {
+			System.out.println("삭제 실패패");
+		}
 		
-		dao = new MemberDAO();
-		MemberDTO dto = dao.selectView(id);
-		int result = dao.deletePost(id);	//게시물 삭제
-		dao.close();
-		
-		alertLocation(resp, "삭제되었습니다", "../member/list.do");
+		alertLocation(resp, " ", "../member/list.do");
 		
 	}
